Update dtype parameter in RetinaConnectionLayer constructor

+++ adult_models.py
@@ -63,7 +63,9 @@ class FullAdultModel(nn.Module):
         super(FullAdultModel, self).__init__()
 
         self.sparse_layout = sparse_layout
-        self.retina_connection = RetinaConnectionLayer(cell_type_indices, 1, 1)
+        self.retina_connection = RetinaConnectionLayer(
+            cell_type_indices, 1, 1, dtype=dtype
+            )     
 
         self.connectome = AdultConnectome(
             adjacency_matrix, layer_number, log_transform_weights, dtype, device, sparse_layout=sparse_layout
